Moves converter classes back to ParquetConverter.scala
